+++ test/views/search.js
@@ -1,7 +1,16 @@
+import { html } from '@aegisjsproject/core/parsers/html.js';
+
 export default (params) => {
-	const pre = document.createElement('pre');
-	const code = document.createElement('code');
-	code.textContent = JSON.stringify(params, null, 4);
-	pre.append(code);
-	return pre;
+	const frag = html`
+		<form action="/search" id="search" action="/search" method="GET">
+			<label for="query">Query</label>
+			<input type="search" name="q" placehoder="Search for..." required="" />
+			<button type="submit" class="btn btn-primary">Search</button>
+		</form>
+		<pre><code>${JSON.stringify(params, null, 4)}</code></pre>
+	`;
+
+	// Cannot set `action` in parsing HTML because of sanitizer
+	frag.getElementById('search').action = '/search';
+	return frag;
 };